requirements.txt
playwright==1.35.0
beautifulsoup4==4.12.2
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
asyncio==3.4.3
argparse==1.4.0
supabase==1.0.3
python-dotenv==1.0.0
fastapi==0.95.1
uvicorn==0.22.0
